+++ truffle/src/com.oracle.truffle.dsl.processor/src/com/oracle/truffle/dsl/processor/library/ExportsGenerator.java
@@ -942,8 +942,7 @@ public class ExportsGenerator extends CodeTypeElementFactory<ExportsData> {
         builder.string(" || ");
         builder.staticReference(dispatchLibraryConstant).string(
                         ".getUncached().dispatch(receiver) == null : ").doubleQuote(
-                                        "Invalid library export '" + libraryExports.getTemplateType().getQualifiedName().toString() +
-                                                        "'. Exported receiver with dynamic dispatch found but not expected.");
+                                        "Invalid library export. Exported receiver with dynamic dispatch found but not expected.");
         builder.end();
         return builder.build();
     }