+++ truffle/src/com.oracle.truffle.object.basic.test/src/com/oracle/truffle/object/basic/test/DeclaredLocationTest.java
@@ -40,10 +40,17 @@
  */
 package com.oracle.truffle.object.basic.test;
 
+import java.util.Arrays;
+import java.util.List;
+
 import org.junit.Assert;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.junit.runners.Parameterized.Parameters;
 
 import com.oracle.truffle.api.object.DynamicObject;
+import com.oracle.truffle.api.object.DynamicObjectLibrary;
 import com.oracle.truffle.api.object.FinalLocationException;
 import com.oracle.truffle.api.object.IncompatibleLocationException;
 import com.oracle.truffle.api.object.Layout;
@@ -51,11 +58,17 @@ import com.oracle.truffle.api.object.Location;
 import com.oracle.truffle.api.object.ObjectType;
 import com.oracle.truffle.api.object.Property;
 import com.oracle.truffle.api.object.Shape;
-import com.oracle.truffle.object.basic.DefaultLayoutFactory;
+import com.oracle.truffle.api.test.AbstractParametrizedLibraryTest;
 
-public class DeclaredLocationTest {
+@RunWith(Parameterized.class)
+public class DeclaredLocationTest extends AbstractParametrizedLibraryTest {
 
-    final Layout layout = new DefaultLayoutFactory().createLayout(Layout.newLayout());
+    @Parameters(name = "{0}")
+    public static List<TestRun> data() {
+        return Arrays.asList(TestRun.values());
+    }
+
+    final Layout layout = Layout.newLayout().build();
     final Shape rootShape = layout.createShape(new ObjectType());
     final Object value = new Object();
     final Location declaredLocation = rootShape.allocator().declaredLocation(value);
@@ -64,9 +77,12 @@ public class DeclaredLocationTest {
     @Test
     public void testDeclaredLocation() {
         DynamicObject object = shapeWithDeclared.newInstance();
-        Assert.assertSame(value, object.get("declared"));
 
-        object.set("declared", value);
+        DynamicObjectLibrary library = createLibrary(DynamicObjectLibrary.class, object);
+
+        Assert.assertSame(value, library.getOrDefault(object, "declared", null));
+
+        library.putIfPresent(object, "declared", value);
         Assert.assertSame(shapeWithDeclared, object.getShape());
 
         Property property = object.getShape().getProperty("declared");
@@ -88,29 +104,36 @@ public class DeclaredLocationTest {
             Assert.assertTrue(e instanceof FinalLocationException);
         }
 
-        Assert.assertSame(value, object.get("declared"));
+        Assert.assertSame(value, library.getOrDefault(object, "declared", null));
     }
 
     @Test
     public void testMigrateDeclaredLocation() {
         DynamicObject object = shapeWithDeclared.newInstance();
+
+        DynamicObjectLibrary library = createLibrary(DynamicObjectLibrary.class, object);
+
         Assert.assertSame(shapeWithDeclared, object.getShape());
-        Assert.assertSame(value, object.get("declared"));
+        Assert.assertSame(value, library.getOrDefault(object, "declared", null));
 
         Object newValue = new Object();
-        object.set("declared", newValue);
+        library.putIfPresent(object, "declared", newValue);
         Assert.assertNotSame(shapeWithDeclared, object.getShape());
-        Assert.assertSame(newValue, object.get("declared"));
+        Assert.assertSame(newValue, library.getOrDefault(object, "declared", null));
     }
 
+    @SuppressWarnings("deprecation")
     @Test
     public void testAddDeclaredLocation() {
         Property property = shapeWithDeclared.getProperty("declared");
 
         DynamicObject object = rootShape.newInstance();
+
+        DynamicObjectLibrary library = createLibrary(DynamicObjectLibrary.class, object);
+
         property.setSafe(object, value, rootShape, shapeWithDeclared);
         Assert.assertSame(shapeWithDeclared, object.getShape());
-        Assert.assertSame(value, object.get("declared"));
+        Assert.assertSame(value, library.getOrDefault(object, "declared", null));
 
         DynamicObject object2 = rootShape.newInstance();
         Object newValue = new Object();
@@ -123,7 +146,7 @@ public class DeclaredLocationTest {
             // Expected
         }
         Assert.assertSame(rootShape, object2.getShape());
-        Assert.assertEquals(false, object2.containsKey("declared"));
+        Assert.assertEquals(false, library.containsKey(object2, "declared"));
     }
 
 }