+++ truffle/src/com.oracle.truffle.object/src/com/oracle/truffle/object/Locations.java
@@ -111,12 +111,12 @@ public abstract class Locations {
         @SuppressWarnings("deprecation")
         @Override
         public boolean canStore(Object val) {
-            return valueEquals(this.value, val);
+            return CoreLocation.valueEquals(this.value, val);
         }
 
         /** @since 0.17 or earlier */
         @Override
-        public final void setInternal(DynamicObject store, Object value) throws IncompatibleLocationException {
+        public final void setInternal(DynamicObject store, Object value, boolean condition) throws IncompatibleLocationException {
             if (!canStore(value)) {
                 CompilerDirectives.transferToInterpreterAndInvalidate();
                 throw new UnsupportedOperationException();